move fake block package
