+++ src/main/kotlin/one/oktw/galaxy/internal/register/RecipeRegister.kt
@@ -1,8 +1,8 @@
 package one.oktw.galaxy.internal.register
 
 import one.oktw.galaxy.Main.Companion.main
-import one.oktw.galaxy.block.FakeBlockItem
-import one.oktw.galaxy.block.FakeBlocks.HT_CRAFTING_TABLE
+import one.oktw.galaxy.block.data.FakeBlockItem
+import one.oktw.galaxy.block.enums.FakeBlocks.HT_CRAFTING_TABLE
 import one.oktw.galaxy.item.enums.ToolType.WRENCH
 import one.oktw.galaxy.item.enums.UpgradeType.*
 import one.oktw.galaxy.item.type.Tool
